refactor more code
