+++ src/components/core/queryForm/QueryForm.css
